chore: enable publishing of independent alpha modules
This is the (hopefully!) final commit to enable us to independently publish the
experimental and dev-preview (collectively "alpha") modules as independent
packages.

After this change, the next automated v2 release will publish both aws-cdk-lib,
as well as one @aws-cdk/aws-foobar-alpha module for each service that is
either in experimental or dev-preview.

NOTE: This is a backport of #16674, minus the changes to `version.v2.json`.

fixes #16432